rethink/remade of items logic, to avoid depending on list.length for critical operations

+++ 11-clase09/routes/endpoints.js
@@ -3,14 +3,15 @@ import Items from '../Items.js'
 
 const endpoints = express.Router();
 
+// const productos = new Items();
 const productos = new Items("Escuadra", "123.45", "https://cdn3.iconfinder.com/data/icons/education-209/64/ruler-triangle-stationary-school-256.png");
-const item2 = {title: "Calculadora", price: "234.56", thumbnail: "https://cdn3.]iconfinder.com/data/icons/education-209/64/calculator-math-tool-school-512.png"}
+const item2 = {title: "Calculadora", price: "234.56", thumbnail: "https://cdn3.iconfinder.com/data/icons/education-209/64/calculator-math-tool-school-512.png"}
 productos.addItem(item2);
 
 const validateId = (id) => {
     return !isNaN(id) ?
         id >= 0 ?
-            id < productos.getItems().length ?
+            id < productos.getLastItemId() ?
                 true
             : `Error: no se encontro ningun producto con ID "${id}"!`
         : "Error: ID de producto no puede ser < 0!"
@@ -18,8 +19,13 @@ const validateId = (id) => {
 }
 
 endpoints.get("/", (req, res) => {
-    res.json({
-        result: "Endpoint inicial!",
+    // res.json({
+    //     result: "Endpoint inicial!",
+    // });
+    const getProds = () => {return {...productos}};
+    res.render('main', {
+        "productos": getProds(),
+        layout: "index",
     });
 });
 
@@ -55,7 +61,7 @@ endpoints.post("/productos/guardar/", (req, res) => {
         productos.addItem(itemData);
         res.json({
             result: "Exito!",
-            id: productos.getItems().length - 1,
+            id: productos.getLastItemId() - 1,
             newProdcut: itemData,
         })
     }else{
@@ -101,13 +107,13 @@ endpoints.put("/productos/actualizar/:id", (req, res) => {
 endpoints.delete("/productos/borrar/:id", (req, res) => {
     const id = Number(req.params.id);
     const result = validateId(id);
-    if (typeof(result) !== "string"){
+    if (typeof(result) !== "string" && typeof(productos.getItem(id)) !== "string"){ // tanto el resultado como el lookup del item son NO string, quiere decir que alguno de los 2 (o ambos) esta/n correcto/s
         res.json({
             result: productos.eliminarItem(id),
         });
     }else{
         res.json({
-            result: result,
+            result: result === true ? productos.getItem(id) : result, // si result tiene algun "insight" ademas de true, imprimirlo, si no, el output de llamar a getItem con id no encontrado!
         });
     }
 });
